// backend/src/enums/publish-state.enum.ts
/**
 * PublishStateEnum
 */
export enum PublishStateEnum {
    USER_CREATED = 'userCreated',
    USER_UPDATED = 'userUpdated',
    USER_GROUP_CREATED = 'userGroupCreated',
    USER_GROUP_UPDATED = 'userGroupUpdated',
    MESSAGE_CREATED = 'messageCreated',
    MESSAGE_TYPE_CREATED = 'messageTypeCreated',
    MESSAGE_TYPE_UPDATED = 'messageTypeUpdated',
    JOB_MESSAGE_STATUS = 'jobSendMessageStatus'
}